backend_metrosence/pyproject.toml
[project]
name = "backend-metrosence"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi>=0.116.1",
    "passlib>=1.7.4",
    "psycopg2>=2.9.10",
    "pydantic[email]>=2.11.7",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.35.0",
]
